Trait ink_env::arithmetic::BaseArithmetic

pub trait BaseArithmetic: Sized + From<u8> + Bounded + Ord + PartialOrd<Self> + Zero + One + Bounded + Add<Self, Output = Self> + AddAssign<Self> + Sub<Self, Output = Self> + SubAssign<Self> + Mul<Self, Output = Self> + MulAssign<Self> + Div<Self, Output = Self> + DivAssign<Self> + CheckedMul + Saturating + TryFrom<u16> + TryFrom<u32> + TryFrom<u64> + TryFrom<u128> + TryFrom<usize> + TryInto<u16> + TryInto<u32> + TryInto<u64> + TryInto<u128> + TryInto<usize> { }

Types that allow for simple arithmetic operations.

Subset of all trait bounds copied over from what Substrate defines for its BaseArithmetic types. We can extend this in the future if needed.