Trait ink_env::backend::EnvBackend

pub trait EnvBackend {
    pub fn set_contract_storage<V>(&mut self, key: &Key, value: &V)
    where
        V: Encode;

    pub fn get_contract_storage<R>(&mut self, key: &Key) -> Result<Option<R>>
    where
        R: Decode;

    pub fn clear_contract_storage(&mut self, key: &Key);

    pub fn decode_input<T>(&mut self) -> Result<T>
    where
        T: Decode;

    pub fn return_value<R>(&mut self, flags: ReturnFlags, return_value: &R) -> !
    where
        R: Encode;

    pub fn println(&mut self, content: &str);

    pub fn hash_bytes<H>(
        &mut self, 
        input: &[u8], 
        output: &mut <H as HashOutput>::Type
    )
    where
        H: CryptoHash;

    pub fn hash_encoded<H, T>(
        &mut self, 
        input: &T, 
        output: &mut <H as HashOutput>::Type
    )
    where
        H: CryptoHash,
        T: Encode;
}

Environmental contract functionality that does not require Environment.

Required methods

pub fn set_contract_storage<V>(&mut self, key: &Key, value: &V) where
V: Encode,

Writes the value to the contract storage under the given key.

pub fn get_contract_storage<R>(&mut self, key: &Key) -> Result<Option<R>> where
R: Decode,

Returns the value stored under the given key in the contract's storage if any.

Errors

If the decoding of the typed value failed

pub fn clear_contract_storage(&mut self, key: &Key)

Clears the contract's storage key entry.

pub fn decode_input<T>(&mut self) -> Result<T> where
T: Decode,

Returns the execution input to the executed contract and decodes it as T.

Note

The input is the 4-bytes selector followed by the arguments of the called function in their SCALE encoded representation.

No prior interaction with the environment must take place before calling this procedure.

Usage

Normally contracts define their own enum dispatch types respective to their exported contructors and messages that implement scale::Decode according to the contructors or messages selectors and their arguments. These enum dispatch types are then given to this procedure as the T.

When using ink! users do not have to construct those enum dispatch types themselves as they are normally generated by the ink! code generation automatically.

Errors

If the given T cannot be properly decoded from the expected input.

pub fn return_value<R>(&mut self, flags: ReturnFlags, return_value: &R) -> ! where
R: Encode,

Returns the value back to the caller of the executed contract.

Note

Calling this method will end contract execution immediately. It will return the given return value back to its caller.

The flags parameter can be used to revert the state changes of the entire execution if necessary.

pub fn println(&mut self, content: &str)

Prints the given contents to the console log.

pub fn hash_bytes<H>(
&mut self,
input: &[u8],
output: &mut <H as HashOutput>::Type
) where
H: CryptoHash,

Conducts the crypto hash of the given input and stores the result in output.

pub fn hash_encoded<H, T>(
&mut self,
input: &T,
output: &mut <H as HashOutput>::Type
) where
H: CryptoHash,
T: Encode,

Conducts the crypto hash of the given encoded input and stores the result in output.

Loading content...