Trait ink_env::backend::TypedEnvBackend

pub trait TypedEnvBackend: EnvBackend {
    pub fn caller<T: Environment>(&mut self) -> Result<T::AccountId>;

    pub fn transferred_balance<T: Environment>(&mut self) -> Result<T::Balance>;

    pub fn weight_to_fee<T: Environment>(
        &mut self, 
        gas: u64
    ) -> Result<T::Balance>;

    pub fn gas_left<T: Environment>(&mut self) -> Result<T::Balance>;

    pub fn block_timestamp<T: Environment>(&mut self) -> Result<T::Timestamp>;

    pub fn account_id<T: Environment>(&mut self) -> Result<T::AccountId>;

    pub fn balance<T: Environment>(&mut self) -> Result<T::Balance>;

    pub fn rent_allowance<T: Environment>(&mut self) -> Result<T::Balance>;

    pub fn block_number<T: Environment>(&mut self) -> Result<T::BlockNumber>;

    pub fn minimum_balance<T: Environment>(&mut self) -> Result<T::Balance>;

    pub fn tombstone_deposit<T: Environment>(&mut self) -> Result<T::Balance>;

    pub fn emit_event<T, Event>(&mut self, event: Event)
    where
        T: Environment,
        Event: Topics + Encode;

    pub fn set_rent_allowance<T>(&mut self, new_value: T::Balance)
    where
        T: Environment;

    pub fn invoke_contract<T, Args>(
        &mut self, 
        call_data: &CallParams<T, Args, ()>
    ) -> Result<()>
    where
        T: Environment,
        Args: Encode;

    pub fn eval_contract<T, Args, R>(
        &mut self, 
        call_data: &CallParams<T, Args, ReturnType<R>>
    ) -> Result<R>
    where
        T: Environment,
        Args: Encode,
        R: Decode;

    pub fn instantiate_contract<T, Args, C>(
        &mut self, 
        params: &CreateParams<T, Args, C>
    ) -> Result<T::AccountId>
    where
        T: Environment,
        Args: Encode;

    pub fn restore_contract<T>(
        &mut self, 
        account_id: T::AccountId, 
        code_hash: T::Hash, 
        rent_allowance: T::Balance, 
        filtered_keys: &[Key]
    )
    where
        T: Environment;

    pub fn terminate_contract<T>(&mut self, beneficiary: T::AccountId) -> !
    where
        T: Environment;

    pub fn transfer<T>(
        &mut self, 
        destination: T::AccountId, 
        value: T::Balance
    ) -> Result<()>
    where
        T: Environment;

    pub fn random<T>(&mut self, subject: &[u8]) -> Result<T::Hash>
    where
        T: Environment;
}

Environmental contract functionality.

Required methods

pub fn caller<T: Environment>(&mut self) -> Result<T::AccountId>

Returns the address of the caller of the executed contract.

Note

For more details visit: [ink_env::caller]

pub fn transferred_balance<T: Environment>(&mut self) -> Result<T::Balance>

Returns the transferred balance for the contract execution.

Note

For more details visit: [ink_env::transferred_balance]

pub fn weight_to_fee<T: Environment>(&mut self, gas: u64) -> Result<T::Balance>

Returns the price for the specified amount of gas.

Note

For more details visit: [ink_env::gas_price]

pub fn gas_left<T: Environment>(&mut self) -> Result<T::Balance>

Returns the amount of gas left for the contract execution.

Note

For more details visit: [ink_env::gas_left]

pub fn block_timestamp<T: Environment>(&mut self) -> Result<T::Timestamp>

Returns the timestamp of the current block.

Note

For more details visit: [ink_env::block_timestamp]

pub fn account_id<T: Environment>(&mut self) -> Result<T::AccountId>

Returns the address of the executed contract.

Note

For more details visit: [ink_env::account_id]

pub fn balance<T: Environment>(&mut self) -> Result<T::Balance>

Returns the balance of the executed contract.

Note

For more details visit: [ink_env::balance]

pub fn rent_allowance<T: Environment>(&mut self) -> Result<T::Balance>

Returns the current rent allowance for the executed contract.

Note

For more details visit: [ink_env::rent_allowance]

pub fn block_number<T: Environment>(&mut self) -> Result<T::BlockNumber>

Returns the current block number.

Note

For more details visit: [ink_env::block_number]

pub fn minimum_balance<T: Environment>(&mut self) -> Result<T::Balance>

Returns the minimum balance that is required for creating an account.

Note

For more details visit: [ink_env::minimum_balance]

pub fn tombstone_deposit<T: Environment>(&mut self) -> Result<T::Balance>

Returns the tombstone deposit of the contract chain.

Note

For more details visit: [ink_env::tombstone_deposit]

pub fn emit_event<T, Event>(&mut self, event: Event) where
T: Environment,
Event: Topics + Encode,

Emits an event with the given event data.

Note

For more details visit: [ink_env::emit_event]

pub fn set_rent_allowance<T>(&mut self, new_value: T::Balance) where
T: Environment,

Sets the rent allowance of the executed contract to the new value.

Note

For more details visit: [ink_env::set_rent_allowance]

pub fn invoke_contract<T, Args>(
&mut self,
call_data: &CallParams<T, Args, ()>
) -> Result<()> where
T: Environment,
Args: Encode,

Invokes a contract message.

Note

For more details visit: [ink_env::invoke_contract]

pub fn eval_contract<T, Args, R>(
&mut self,
call_data: &CallParams<T, Args, ReturnType<R>>
) -> Result<R> where
T: Environment,
Args: Encode,
R: Decode,

Evaluates a contract message and returns its result.

Note

For more details visit: [ink_env::eval_contract]

pub fn instantiate_contract<T, Args, C>(
&mut self,
params: &CreateParams<T, Args, C>
) -> Result<T::AccountId> where
T: Environment,
Args: Encode,

Instantiates another contract.

Note

For more details visit: [ink_env::instantiate_contract]

pub fn restore_contract<T>(
&mut self,
account_id: T::AccountId,
code_hash: T::Hash,
rent_allowance: T::Balance,
filtered_keys: &[Key]
) where
T: Environment,

Restores a smart contract tombstone.

Note

For more details visit: [ink_env::restore_contract]

pub fn terminate_contract<T>(&mut self, beneficiary: T::AccountId) -> ! where
T: Environment,

Terminates a smart contract.

Note

For more details visit: [ink_env::terminate_contract]

pub fn transfer<T>(
&mut self,
destination: T::AccountId,
value: T::Balance
) -> Result<()> where
T: Environment,

Transfers value from the contract to the destination account ID.

Note

For more details visit: [ink_env::transfer]

pub fn random<T>(&mut self, subject: &[u8]) -> Result<T::Hash> where
T: Environment,

Returns a random hash seed.

Note

For more details visit: [ink_env::random]

Loading content...