ink_env

Module call

Utilities to call or instantiate contracts on the chain.

Modules

state

Type states that tell what state of a instantiation argument has not yet been set properly for a valid construction.

utils

Utility types for the cross-contract calling API.

Structs

Call

The default call type for cross-contract calls, for calling into the latest call host function. This adds the additional weight limit parameter proof_size_limit as well as storage_deposit_limit.

CallBuilder

Builds up a cross contract call.

CallParams

The final parameters to the cross-contract call.

CreateBuilder

Builds up contract instantiations.

CreateParams

Builds up contract instantiations.

DelegateCall

The delegatecall call type. Performs a call with the given code hash.

Execution

The input data and the expected return type of a contract execution.

ExecutionInput

The input data for a smart contract execution.

LimitParamsV2

Defines the limit params for the new ext::instantiate host function. todo: rename

Selector

The function selector.

Traits

ConstructorReturnType

Represents any type that can be returned from an ink! constructor. The following contract implements the four different return type signatures implementing this trait:

Executor

Implemented in different environments to perform contract execution.

FromAddr

Contracts that can be constructed from an AccountId.

Functions

build_call

Returns a new CallBuilder to build up the parameters to a cross-contract call that uses the “default” ABI for calls for the ink! project.

build_call_ink

Returns a new CallBuilder to build up the parameters to a cross-contract call that uses ink! ABI Encoding (i.e. with SCALE codec for input/output encode/decode).

build_call_sol

Returns a new CallBuilder to build up the parameters to a cross-contract call that uses Solidity ABI Encoding.

build_create

Returns a new CreateBuilder to build up the parameters to a cross-contract instantiation that uses the “default” ABI for calls for the ink! project.

build_create_ink

Returns a new CreateBuilder to build up the parameters to a cross-contract instantiation that uses ink! ABI Encoding (i.e. with SCALE codec for input/output encode/decode).

build_create_sol

Returns a new CreateBuilder to build up the parameters to a cross-contract instantiation that uses Solidity ABI Encoding.