Struct ink_env::call::CreateBuilder

pub struct CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Args, Salt, RetType>
where
    E: Environment,
{ /* private fields */ }

Builds up contract instantiations.

Implementations

impl<E, ContractRef, Limits, Endowment, Args, Salt, RetType> CreateBuilder<E, ContractRef, Unset<E::Hash>, Limits, Endowment, Args, Salt, RetType>
where E: Environment,

pub fn code_hash( self, code_hash: E::Hash, ) -> CreateBuilder<E, ContractRef, Set<E::Hash>, Limits, Endowment, Args, Salt, RetType>

Sets the used code hash for the contract instantiation.

impl<E, ContractRef, CodeHash, Endowment, Args, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, Set<LimitParamsV1>, Endowment, Args, Salt, RetType>
where E: Environment,

pub fn gas_limit(self, gas_limit: u64) -> Self

Sets the maximum allowed gas costs for the contract instantiation.

impl<E, ContractRef, CodeHash, Endowment, Args, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, Set<LimitParamsV2<E>>, Endowment, Args, Salt, RetType>
where E: Environment,

pub fn instantiate_v1( self, ) -> CreateBuilder<E, ContractRef, CodeHash, Set<LimitParamsV1>, Endowment, Args, Salt, RetType>

Switch to the original instantiate host function API, which only allows the gas_limit limit parameter (equivalent to the ref_time_limit in the latest instantiate_v2).

This method instance is used to allow usage of the generated builder methods for constructors which initialize the builder with the new LimitParamsV2 type.

pub fn ref_time_limit(self, ref_time_limit: u64) -> Self

Sets the ref_time_limit part of the weight limit for the contract instantiation.

pub fn proof_size_limit(self, proof_size_limit: u64) -> Self

Sets the proof_size_limit part of the weight limit for the contract instantiation.

pub fn storage_deposit_limit(self, storage_deposit_limit: E::Balance) -> Self

Sets the storage_deposit_limit for the contract instantiation.

impl<E, ContractRef, CodeHash, Limits, Args, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, Limits, Unset<E::Balance>, Args, Salt, RetType>
where E: Environment,

pub fn endowment( self, endowment: E::Balance, ) -> CreateBuilder<E, ContractRef, CodeHash, Limits, Set<E::Balance>, Args, Salt, RetType>

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, CodeHash, Limits, Endowment, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Unset<ExecutionInput<EmptyArgumentList>>, Salt, RetType>
where E: Environment,

pub fn exec_input<Args>( self, exec_input: ExecutionInput<Args>, ) -> CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Set<ExecutionInput<Args>>, Salt, RetType>

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, CodeHash, Limits, Endowment, Args, RetType> CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Args, Unset<Salt>, RetType>
where E: Environment,

pub fn salt_bytes<Salt>( self, salt: Salt, ) -> CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Args, Set<Salt>, RetType>
where Salt: AsRef<[u8]>,

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, CodeHash, Limits, Endowment, Args, Salt> CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Args, Salt, Unset<ReturnType<()>>>
where E: Environment,

pub fn returns<R>( self, ) -> CreateBuilder<E, ContractRef, CodeHash, Limits, Endowment, Args, Salt, Set<ReturnType<R>>>
where ContractRef: FromAccountId<E>, R: ConstructorReturnType<ContractRef>,

Sets the type of the returned value upon the execution of the constructor.

Note

Constructors are not able to return arbitrary values. Instead a successful call to a constructor returns the address at which the contract was instantiated.

Therefore this must always be a reference (i.e ContractRef) to the contract you’re trying to instantiate.

impl<E, ContractRef, Limits, Args, Salt, RetType> CreateBuilder<E, ContractRef, Set<E::Hash>, Set<Limits>, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, Set<ReturnType<RetType>>>
where E: Environment,

pub fn params(self) -> CreateParams<E, ContractRef, Limits, Args, Salt, RetType>

Finalizes the create builder, allowing it to instantiate a contract.

impl<E, ContractRef, Args, Salt, RetType> CreateBuilder<E, ContractRef, Set<E::Hash>, Set<LimitParamsV2<E>>, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, Set<ReturnType<RetType>>>
where E: Environment, ContractRef: FromAccountId<E>, Args: Encode, Salt: AsRef<[u8]>, RetType: ConstructorReturnType<ContractRef>,

pub fn instantiate( self, ) -> <RetType as ConstructorReturnType<ContractRef>>::Output

Instantiates the contract and returns its account ID back to the caller.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_instantiate method instead.

pub fn try_instantiate( self, ) -> Result<ConstructorResult<<RetType as ConstructorReturnType<ContractRef>>::Output>, Error>

Instantiates the contract and returns its account ID back to the caller.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E, ContractRef, Args, Salt, RetType> CreateBuilder<E, ContractRef, Set<E::Hash>, Set<LimitParamsV1>, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, Set<ReturnType<RetType>>>
where E: Environment, ContractRef: FromAccountId<E>, Args: Encode, Salt: AsRef<[u8]>, RetType: ConstructorReturnType<ContractRef>,

pub fn instantiate( self, ) -> <RetType as ConstructorReturnType<ContractRef>>::Output

Instantiates the contract and returns its account ID back to the caller.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_instantiate method instead.

pub fn try_instantiate( self, ) -> Result<ConstructorResult<<RetType as ConstructorReturnType<ContractRef>>::Output>, Error>

Instantiates the contract and returns its account ID back to the caller.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.