Struct ink_env::call::CreateParams

pub struct CreateParams<E, ContractRef, Limits, Args, Salt, R>
where
    E: Environment,
{ /* private fields */ }

Builds up contract instantiations.

Implementations

impl<E, ContractRef, Limits, Args, Salt, R> CreateParams<E, ContractRef, Limits, Args, Salt, R>
where E: Environment,

pub fn code_hash(&self) -> &E::Hash

The code hash of the contract.

pub fn endowment(&self) -> &E::Balance

The endowment for the instantiated contract.

pub fn exec_input(&self) -> &ExecutionInput<Args>

The raw encoded input data.

pub fn update_selector(&mut self, selector: Selector)

Modify the selector.

Useful when using the CreateParams generated as part of the ContractRef, but using a custom selector.

impl<E, ContractRef, Args, Salt, R> CreateParams<E, ContractRef, LimitParamsV2<E>, Args, Salt, R>
where E: Environment,

pub fn ref_time_limit(&self) -> u64

Gets the ref_time_limit part of the weight limit for the contract instantiation.

pub fn proof_size_limit(&self) -> u64

Gets the proof_size_limit part of the weight limit for the contract instantiation.

pub fn storage_deposit_limit(&self) -> Option<&E::Balance>

Gets the storage_deposit_limit for the contract instantiation.

impl<E, ContractRef, Args, Salt, R> CreateParams<E, ContractRef, LimitParamsV1, Args, Salt, R>
where E: Environment,

pub fn gas_limit(&self) -> u64

The gas limit for the contract instantiation.

impl<E, ContractRef, Limits, Args, Salt, R> CreateParams<E, ContractRef, Limits, Args, Salt, R>
where E: Environment, Salt: AsRef<[u8]>,

pub fn salt_bytes(&self) -> &Salt

The salt for determining the hash for the contract account ID.

impl<E, ContractRef, Args, Salt, R> CreateParams<E, ContractRef, LimitParamsV2<E>, Args, Salt, R>
where E: Environment, ContractRef: FromAccountId<E>, Args: Encode, Salt: AsRef<[u8]>, R: ConstructorReturnType<ContractRef>,

pub fn instantiate(&self) -> <R as ConstructorReturnType<ContractRef>>::Output

Instantiates the contract and returns its account ID back to the caller.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_instantiate method instead.

pub fn try_instantiate( &self, ) -> Result<ConstructorResult<<R as ConstructorReturnType<ContractRef>>::Output>, Error>

Instantiates the contract and returns its account ID back to the caller.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E, ContractRef, Args, Salt, R> CreateParams<E, ContractRef, LimitParamsV1, Args, Salt, R>
where E: Environment, ContractRef: FromAccountId<E>, Args: Encode, Salt: AsRef<[u8]>, R: ConstructorReturnType<ContractRef>,

pub fn instantiate(&self) -> <R as ConstructorReturnType<ContractRef>>::Output

Instantiates the contract and returns its account ID back to the caller.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_instantiate method instead.

pub fn try_instantiate( &self, ) -> Result<ConstructorResult<<R as ConstructorReturnType<ContractRef>>::Output>, Error>

Instantiates the contract and returns its account ID back to the caller.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.