ink_env::call

Trait ConstructorReturnType

pub trait ConstructorReturnType<C, Abi> {
    type Output;
    type Error: DecodeConstructorError<Abi>;

    const IS_RESULT: bool = false;

    // Required method
    fn ok(value: C) -> Self::Output;

    // Provided method
    fn err(_err: Self::Error) -> Option<Self::Output> { ... }
}

Represents any type that can be returned from an ink! constructor. The following contract implements the four different return type signatures implementing this trait:

Self

Result<Self, Error>

Contract

Result<Contract, Error>

#[ink::contract]
mod contract {
    #[ink(storage)]
    pub struct Contract {}

    #[derive(Debug, PartialEq, Eq)]
    #[ink::scale_derive(Encode, Decode, TypeInfo)]
    pub enum Error {
        Foo,
    }

    impl Contract {
        #[ink(constructor)]
        pub fn new_self() -> Self {
            Self {}
        }

        #[ink(constructor)]
        pub fn new_storage_name() -> Contract {
            Contract {}
        }

        #[ink(constructor)]
        pub fn new_result_self() -> Result<Self, Error> {
            Ok(Self {})
        }

        #[ink(constructor)]
        pub fn new_result_storage_name() -> Result<Contract, Error> {
            Ok(Contract {})
        }

        #[ink(message)]
        pub fn message(&self) {}
    }
}

These constructor return signatures are then used by the ContractRef codegen for the CreateBuilder::returns type parameter.

Provided Associated Constants

const IS_RESULT: bool = false

Is true if Self is Result<C, E>.

Required Associated Types

type Output

The actual return type of the constructor.

If a constructor returns Self, then Output = Self

If a constructor returns a Result<Self, E>, then Output = Result<Self, E>

type Error: DecodeConstructorError<Abi>

The error type of the constructor return type.

Required Methods

fn ok(value: C) -> Self::Output

Construct a success value of the Output type.

Provided Methods

fn err(_err: Self::Error) -> Option<Self::Output>

Construct an error value of the Output type.

Result implementations should return Some(Err(err)), otherwise default to None.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.