Trait ink_env::call::Executor

pub trait Executor<E: Environment> {
    type Error;

    // Required method
    fn exec<Args, Output>(
        &self,
        input: &ExecutionInput<Args>,
    ) -> Result<MessageResult<Output>, Self::Error>
       where Args: Encode,
             Output: Decode;
}

Implemented in different environments to perform contract execution.

Required Associated Types

type Error

The type of the error that can be returned during execution.

Required Methods

fn exec<Args, Output>( &self, input: &ExecutionInput<Args>, ) -> Result<MessageResult<Output>, Self::Error>
where Args: Encode, Output: Decode,

Perform the contract execution with the given input data, and return the result.

Object Safety

This trait is not object safe.