Struct ink_env::call::utils::ArgumentList

pub struct ArgumentList<Head, Rest> { /* fields omitted */ }

An argument list.

This type is constructed mainly at compile type via type constructions to avoid having to allocate heap memory while constructing the encoded arguments. The potentially heap allocating encoding is done right at the end where we can leverage the static environmental buffer instead of allocating heap memory.

Implementations

impl ArgumentList<ArgumentListEnd, ArgumentListEnd>

pub fn empty() -> EmptyArgumentList

Creates a new empty argument list.

pub fn push_arg<T>(self, arg: T) -> ArgumentList<Argument<T>, Self> where
T: Encode,

Pushes the first argument to the empty argument list.

impl<Head, Rest> ArgumentList<Argument<Head>, Rest>

pub fn push_arg<T>(self, arg: T) -> ArgumentList<Argument<T>, Self> where
T: Encode,

Pushes another argument to the argument list.