ink_env::call::utils

Trait DecodeMessageResult

pub trait DecodeMessageResult<Abi>: Sized {
    // Required method
    fn decode_output(buffer: &[u8]) -> Result<MessageResult<Self>>;
}

A trait for decoding the output of a message based on different ABIs. This is necessary as contracts with different ABIs have different return types. For example, Solidity contracts return the output directly without MessageResult.

Required Methods

fn decode_output(buffer: &[u8]) -> Result<MessageResult<Self>>

Decodes the output of a message call, requiring the output to be wrapped with MessageResult (if not included in the output).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.