Module ink_env::e2e

Module for the logic behind ink!’s End-to-End testing framework.

Re-exports

pub use smart_bench_macro;

pub use env_logger;

pub use tokio;

Structs

Client

The Client takes care of communicating with the node.

PairSigner

A [Signer] implementation that can be constructed from an [Pair].

Enums

AccountKeyring

Set of test accounts.

Error

An error occurred while interacting with the Substrate node.

SubstrateConfig

Default set of commonly used types by Substrate runtimes.

Constants

LOG_PREFIX

This prefix will be used for log output. It is set by each #[ink::e2e_test] with the function name as String. This way it is possible to distinguish the lines in stdout and stderr, to still know which line belongs to which test.

Statics

INIT

We use this to only initialize env_logger once.

Traits

InkConstructor

Trait for contract constructors.

InkMessage

Trait for contract messages.

Functions

alice

Returns the default sp_keyring::sr25519::Keyring for //Alice.

bob

Returns the default sp_keyring::sr25519::Keyring for //Bob.

charlie

Returns the default sp_keyring::sr25519::Keyring for //Charlie.

dave

Returns the default sp_keyring::sr25519::Keyring for //Dave.

eve

Returns the default sp_keyring::sr25519::Keyring for //Eve.

ferdie

Returns the default sp_keyring::sr25519::Keyring for //Ferdie.

log_error

Writes msg to stderr.

log_info

Writes msg to stdout.

log_prefix

Returns the name of the test which is currently executed.

one

Returns the default sp_keyring::sr25519::Keyring for //One.

two

Returns the default sp_keyring::sr25519::Keyring for //Two.

Type Definitions

PolkadotConfig

Default set of commonly used types by Polkadot nodes.

Signer

Signer that is used throughout the E2E testing.