Enum ink_env::e2e::Error

pub enum Error<C, E> where
    C: Config,
    E: Environment,
    <E as Environment>::Balance: Debug, {
    InstantiateDryRun(ContractInstantiateResult<C::AccountId, E::Balance>),
    InstantiateExtrinsic(DispatchError),
    CallDryRun(ContractExecResult<E::Balance>),
    CallExtrinsic(DispatchError),
}

An error occurred while interacting with the Substrate node.

We only convey errors here that are caused by the contract’s testing logic. For anything concerning the node (like inability to communicate with it, fetch the nonce, account info, etc.) we panic.

Variants

InstantiateDryRun(ContractInstantiateResult<C::AccountId, E::Balance>)

The instantiate_with_code dry run failed.

InstantiateExtrinsic(DispatchError)

The instantiate_with_code extrinsic failed.

CallDryRun(ContractExecResult<E::Balance>)

The call dry run failed.

CallExtrinsic(DispatchError)

The call extrinsic failed.