Struct ink_env::engine::off_chain::db::accounts::Account

pub struct Account {
    balance: TypedEncoded<Balance>,
    kind: AccountKind,
}

An account within the chain.

Fields
balance: TypedEncoded<Balance>

The balance of the account.

kind: AccountKind

The kind of the account.

Implementations

impl Account

pub fn balance<T>(&self) -> Result<T::Balance, AccountError> where
T: Environment,

Returns the balance of the account.

pub fn set_balance<T>(
&mut self,
new_balance: T::Balance
) -> Result<(), AccountError> where
T: Environment,

Sets the balance of the account.

pub(in engine::off_chain::db::accounts) fn contract_or_err(
&self
) -> Result<&ContractAccount, AccountError>

Returns the contract account or an error if it is a user account.

pub(in engine::off_chain::db::accounts) fn contract_or_err_mut(
&mut self
) -> Result<&mut ContractAccount, AccountError>

Returns the contract account or an error if it is a user account.

pub fn rent_allowance<T>(&self) -> Result<T::Balance, AccountError> where
T: Environment,

Returns the rent allowance of the contract account or an error.

pub fn set_rent_allowance<T>(
&mut self,
new_rent_allowance: T::Balance
) -> Result<(), AccountError> where
T: Environment,

Sets the rent allowance for the contract account or returns an error.

pub fn set_storage<T>(
&mut self,
at: Key,
new_value: &T
) -> Result<(), AccountError> where
T: Encode,

Sets the contract storage of key to the new value.

pub fn clear_storage(&mut self, at: Key) -> Result<(), AccountError>

Clears the contract storage at key.

pub fn get_storage<T>(&self, at: Key) -> Result<Option<T>, AccountError> where
T: Decode,

Returns the value stored in the contract storage at the given key.

pub fn get_storage_rw(&self) -> Result<(usize, usize), AccountError>

Returns the total number of reads and write from and to the contract's storage.

pub fn count_used_storage_cells(&self) -> Result<usize, AccountError>

Returns the amount of used storage entries.