Struct ink_env::engine::off_chain::db::block::Block

pub struct Block {
    number: TypedEncoded<BlockNumber>,
    timestamp: TypedEncoded<OffTimestamp>,
    entropy: TypedEncoded<Hash>,
}

An emulated block in the chain.

Fields
number: TypedEncoded<BlockNumber>

The current block number.

timestamp: TypedEncoded<OffTimestamp>

The timestamp of the block.

entropy: TypedEncoded<Hash>

The randomization entropy for a block.

Note

Can optionally be set for more control via [crate::set_block_randomization_hash].

Implementations

impl Block

pub fn new<T>(number: T::BlockNumber, timestamp: T::Timestamp) -> Self where
T: Environment,

Creates a new block for the given number and time stamp.

pub fn number<T>(&self) -> Result<T::BlockNumber, OffChainError> where
T: Environment,

Returns the block number.

pub fn timestamp<T>(&self) -> Result<T::Timestamp, OffChainError> where
T: Environment,

Returns the timestamp of the block.

pub fn set_entropy<T>(
&mut self,
new_entropy: T::Hash
) -> Result<(), OffChainError> where
T: Environment,

Sets the entropy of this block to the given entropy.

Note

This is mainly used to control what crate::random returns in the off-chain environment.

pub fn random<T>(&self, subject: &[u8]) -> Result<T::Hash, OffChainError> where
T: Environment,

Returns a randomized hash.

Note

This is the off-chain environment implementation of crate::random. It provides the same behaviour in that it will likely yield the same hash for the same subjects within the same block (or execution context).

Returned hashes on the surface might appear random, however for testability purposes the actual implementation is quite simple and computes those "random" hashes by wrapping XOR of the internal entry hash with the eventually repeated sequence of the subject buffer.