Struct ink_env::engine::off_chain::EnvInstance

pub struct EnvInstance {
    accounts: AccountsDb,
    exec_context: Vec<ExecContext>,
    chain_spec: ChainSpec,
    blocks: Vec<Block>,
    console: Console,
    emitted_events: EmittedEventsRecorder,
    clear_storage_disabled: bool,
}

The off-chain environment.

Mainly used for off-chain testing.

Fields
accounts: AccountsDb

The accounts database of the environment.

exec_context: Vec<ExecContext>

Current execution context and context.

chain_spec: ChainSpec

The general chain spec.

blocks: Vec<Block>

The blocks of the chain.

console: Console

The console to print debug contents.

emitted_events: EmittedEventsRecorder

Emitted events recorder.

clear_storage_disabled: bool

Set to true to disable clearing storage

Implementations

impl EnvInstance

pub(in engine::off_chain::impls) fn callee_account(
&self
) -> &Account

Returns the callee account.

pub(in engine::off_chain::impls) fn callee_account_mut(
&mut self
) -> &mut Account

Returns the callee account as mutable reference.

impl EnvInstance

pub(in engine::off_chain::impls) fn transfer_impl<T>(
&mut self,
destination: &T::AccountId,
value: T::Balance
) -> Result<()> where
T: Environment,

pub(in engine::off_chain::impls) fn terminate_contract_impl<T>(
&mut self,
beneficiary: T::AccountId
) -> ! where
T: Environment,

impl EnvInstance

pub fn uninitialized() -> Self

Creates a new uninitialized off-chain environment.

pub fn is_initialized(&self) -> bool

Returns true if the off-chain environment is uninitialized.

pub fn initialize_or_reset_as_default<T>(&mut self) -> Result<()> where
T: Environment,
<T as Environment>::AccountId: From<[u8; 32]>,

Either resets or initializes the off-chain environment to default values.

pub fn reset(&mut self)

Resets the off-chain environment to unintialized state.

pub fn initialize_as_default<T>(&mut self) -> Result<()> where
T: Environment,
<T as Environment>::AccountId: From<[u8; 32]>,

Initializes the whole off-chain environment.

Note

This is needed since we operate on a static instance that cannot be made generic over all environmental types and thus we are required to initialize it upon program start uninitialized which is why we have TypedEncoded wrappers.

The off-chain environment requires to be initialized before every usage.

This routine implements a default initialization that should be fine for most use cases.

pub fn advance_block<T>(&mut self) -> Result<()> where
T: Environment,

Advances the chain by a single block.

pub(in engine::off_chain) fn exec_context(
&self
) -> Result<&ExecContext, OffChainError>

Returns the current execution context.

pub(in engine::off_chain) fn exec_context_mut(
&mut self
) -> Result<&mut ExecContext, OffChainError>

Returns the current execution context.

pub(in engine::off_chain) fn current_block(
&self
) -> Result<&Block, OffChainError>

Returns the current block of the chain.

pub(in engine::off_chain) fn current_block_mut(
&mut self
) -> Result<&mut Block, OffChainError>

Returns a mutable reference to the current block of the chain.

pub(in engine::off_chain) fn chain_spec_mut(
&mut self
) -> &mut ChainSpec