Struct ink_env::engine::off_chain::typed_encoded::TypedEncoded

pub struct TypedEncoded<T> {
    encoded: Vec<u8>,
    type_id: Option<TypeId>,
    marker: PhantomData<fn() -> T>,
}

A wrapper around an encoded entity that only allows type safe accesses.

Note

Checks are implemented at runtime.

Fields
encoded: Vec<u8>

The bytes of the encoded representation of the type.

type_id: Option<TypeId>

The unique identifier of the encoded type.

Note

If this is None it means that the instance is currently untyped and will take over any given type upon the first typed interaction.

This is needed since instances of TypedEncoded are going to be used in static memory where it isn't possible to decide about the used types given by Environment at initialization.

marker: PhantomData<fn() -> T>

Classification marker.

Note

This shouldn't be the typed that is actually stored as encoded representation in self.encoded but should primarily be an abstract marker type that may be used for classification.

The idea behind the marker is to say that whenever two instances of TypedEncoded share a marker they are guaranteed to also have a common (but unknown) type_id so they can decode to the same original type and thus we can allow to interoperate on them.

Example

The TestEnv might use one abstract marker for every of the fundamental FRAME types: Balance, AccountId, Hash, etc. With this and the explicit guarantee that two instances of TypedEncoded with the same abstract marker also share the same (unknown) type_id it is possible to allow them to interoperate.

Implementations

impl<M> TypedEncoded<M>

pub fn uninitialized() -> Self

Creates a new uninitialized instance.

pub fn new<T>(value: &T) -> Self where
T: Encode + 'static,

Creates a new typed-encoded initialized by value of type T.

pub fn try_initialize<T>(&mut self, value: &T) -> Result<(), TypedEncodedError> where
T: Encode + 'static,

Initializes self with a given encodable value.

Errors

If self has already been initialized or is an initialized instance.

pub fn encoded_bytes(&self) -> Result<&[u8], TypedEncodedError>

Returns the encoded bytes representation.

Errors

If the instance is still uninitialized.

pub fn encoded_bytes_mut(&mut self) -> Result<&mut [u8], TypedEncodedError>

Returns a mutable reference to the encoded bytes representation.

Errors

If the instance is still uninitialized.

pub(in engine::off_chain::typed_encoded) fn type_id(
&self
) -> Result<TypeId, TypedEncodedError>

Returns the type ID if the instance has already been initialized.

Errors

Returns an appropriate error in case the instance is uninitialized.

pub(in engine::off_chain::typed_encoded) fn check_matching_types(
&self,
other: &Self
) -> Result<(), TypedEncodedError>

Returns Ok if both types are encoded with the same type.

pub(in engine::off_chain::typed_encoded) fn check_enforced_type<T>(
&self
) -> Result<(), TypedEncodedError> where
T: 'static,

Returns Ok if T is the type represented by the typed encoded instance.

pub fn decode<T>(&self) -> Result<T, TypedEncodedError> where
T: Decode + 'static,

Decodes the instance.

Note

This effectively creates a clone of the encoded value.

pub fn assign<T>(&mut self, value: &T) -> Result<(), TypedEncodedError> where
T: Encode + 'static,

Assigns the given T to self.

pub fn eval<T, F, R>(&self, other: &Self, f: F) -> Result<R, TypedEncodedError> where
T: Decode + 'static,
F: FnOnce(&T, &T) -> R,

Evaluates the given clousure on the given typed encoded instances.

pub fn eval_mut<T, F, R>(
&mut self,
other: &Self,
f: F
) -> Result<R, TypedEncodedError> where
T: Decode + Encode + 'static,
F: FnOnce(&mut T, &T) -> R,

Evaluates the given clousure on the given typed decoded instances and writes back the result into the typed encoded instance.

pub fn eq<T>(&self, other: &Self) -> Result<bool, TypedEncodedError> where
T: PartialEq + Decode + 'static,

Returns true if both instances are of type T and are equal.

Note

The equality check is performed on decoded instances.

pub fn cmp<T>(&self, other: &Self) -> Result<Ordering, TypedEncodedError> where
T: PartialOrd + Ord + Decode + 'static,

Returns order relation if both instances are of type T.

Note

The order relation is performed on the decoded instances.

pub fn hash<T, H>(&self, state: &mut H) -> Result<(), TypedEncodedError> where
T: Decode + Hash + 'static,
H: Hasher,

Computes the hash of the decoded typed instance if types match.