Enum ink_env::Error

pub enum Error {

    Decode(Error),
    OffChain(OffChainError),
    CalleeTrapped,
    CalleeReverted,
    KeyNotFound,
    _BelowSubsistenceThreshold,
    TransferFailed,
    _EndowmentTooLow,
    CodeNotFound,
    NotCallable,
    Unknown,
    LoggingDisabled,
    EcdsaRecoveryFailed,

}

Errors that can be encountered upon environmental interaction.

Variants

Decode(Error)

Error upon decoding an encoded value.

OffChain(OffChainError)

An error that can only occur in the off-chain environment.

CalleeTrapped

The call to another contract has trapped.

CalleeReverted

The call to another contract has been reverted.

KeyNotFound

The queried contract storage entry is missing.

_BelowSubsistenceThreshold

Deprecated and no longer returned: There is only the minimum balance.

TransferFailed

Transfer failed for other not further specified reason. Most probably reserved or locked balance of the sender that was preventing the transfer.

_EndowmentTooLow

Deprecated and no longer returned: Endowment is no longer required.

CodeNotFound

No code could be found at the supplied code hash.

NotCallable

The account that was called is no contract, but a plain account.

Unknown

An unknown error has occurred.

LoggingDisabled

The call to seal_debug_message had no effect because debug message recording was disabled.

EcdsaRecoveryFailed

ECDSA pubkey recovery failed. Most probably wrong recovery id or signature.