ink_env

Function code_hash

pub fn code_hash(addr: &H160) -> Result<H256>

Retrieves the code hash of the contract at the specified account id.

Errors

If no code hash was found for the specified account id.

If the returned value cannot be properly decoded.