ink_env
Function contains_contract_storage

pub fn contains_contract_storage<K>(key: &K) -> Option<u32>
where
    K: Encode,

Checks whether there is a value stored under the given storage key in the contract’s storage.

If a value is stored under the specified key, the size of the value is returned.