Function ink_env::set_contract_storage

pub fn set_contract_storage<K, V>(key: &K, value: &V) -> Option<u32>where
    K: Encode,
    V: Storable,

Writes the value to the contract storage under the given storage key and returns the size of pre-existing value if any.

Panics

If the encode length of value exceeds the configured maximum value length of a storage entry.