ink_env
Function transfer

pub fn transfer<E>(destination: H160, value: U256) -> Result<()>
where
    E: Environment,

Transfers value from the contract to the destination account ID.

Note

This is more efficient and simpler than the alternative to make a no-op contract call or invoke a runtime function that performs the transaction.

Errors

If the contract does not have sufficient free funds.

If the transfer had brought the sender’s total balance below the minimum balance. You need to use terminate_contract in case this is your intention.