Struct ink_env::RentParams

pub struct RentParams<T: Environment> {
    pub total_balance: T::Balance,
    pub free_balance: T::Balance,
    pub subsistence_threshold: T::Balance,
    pub deposit_per_contract: T::Balance,
    pub deposit_per_storage_byte: T::Balance,
    pub deposit_per_storage_item: T::Balance,
    pub rent_allowance: T::Balance,
    pub rent_fraction: T::RentFraction,
    pub storage_size: u32,
    pub code_size: u32,
    pub code_refcount: u32,
    pub _reserved: Option<()>,
}

Information needed for rent calculations that can be requested by a contract.

Fields
total_balance: T::Balance

The total balance of the contract. Includes the balance transferred from the caller.

free_balance: T::Balance

The free balance of the contract, i.e. the portion of the contract’s balance that is not reserved. Includes the balance transferred from the caller.

subsistence_threshold: T::Balance

Subsistence threshold is the extension of the minimum balance (aka existential deposit) by the tombstone deposit, required for leaving a tombstone.

Rent or any contract initiated balance transfer mechanism cannot make the balance lower than the subsistence threshold in order to guarantee that a tombstone is created.

The only way to completely kill a contract without a tombstone is calling seal_terminate.

deposit_per_contract: T::Balance

The balance every contract needs to deposit to stay alive indefinitely.

This is different from the tombstone deposit because this only needs to be deposited while the contract is alive. Costs for additional storage are added to this base cost.

This is a simple way to ensure that contracts with empty storage eventually get deleted by making them pay rent. This creates an incentive to remove them early in order to save rent.

deposit_per_storage_byte: T::Balance

The balance a contract needs to deposit per storage byte to stay alive indefinitely.

Let’s suppose the deposit is 1,000 BU (balance units)/byte and the rent is 1 BU/byte/day, then a contract with 1,000,000 BU that uses 1,000 bytes of storage would pay no rent. But if the balance reduced to 500,000 BU and the storage stayed the same at 1,000, then it would pay 500 BU/day.

deposit_per_storage_item: T::Balance

The balance a contract needs to deposit per storage item to stay alive indefinitely.

It works as Self::deposit_per_storage_byte but for storage items.

rent_allowance: T::Balance

The contract’s rent allowance, the rent mechanism cannot consume more than this.

rent_fraction: T::RentFraction

The fraction of the deposit costs that should be used as rent per block.

When a contract does not have enough balance deposited to stay alive indefinitely it needs to pay per block for the storage it consumes that is not covered by the deposit. This determines how high this rent payment is per block as a fraction of the deposit costs.

storage_size: u32

The total number of bytes used by this contract.

It is a sum of each key-value pair stored by this contract.

code_size: u32

Sum of instrumented and pristine code length.

code_refcount: u32

The number of contracts using this executable.

_reserved: Option<()>

Reserved for backwards compatible changes to this data structure.