Function ink_env::test::push_execution_context

pub fn push_execution_context<T>(
    caller: T::AccountId, 
    callee: T::AccountId, 
    gas_limit: T::Balance, 
    endowment: T::Balance, 
    call_data: CallData
) where
    T: Environment, 

Pushes a contract execution context.

This is the data behind a single instance of a contract call.

Note

Together with pop_execution_context this can be used to emulated nested calls.