ink_env::test

Function set_clear_storage_disabled

pub fn set_clear_storage_disabled(_disable: bool)

Set to true to disable clearing storage

Note

Useful for benchmarks because it ensures the initialized storage is maintained across runs, because lazy storage structures automatically clear their associated cells when they are dropped.