Trait ink_env::Topics

pub trait Topics {
    type RemainingTopics: EventTopicsAmount;

    fn topics<E, B>(
        &self,
        builder: TopicsBuilder<Uninit, E, B>
    ) -> <B as TopicsBuilderBackend<E>>::Output
    where
        E: Environment,
        B: TopicsBuilderBackend<E>;
}

Implemented by event types to guide the event topic serialization using the topics builder.

Normally this trait should be implemented automatically via the ink! codegen.

Required Associated Types

type RemainingTopics: EventTopicsAmount

Type state indicating how many event topics are to be expected by the topics builder.

Required Methods

fn topics<E, B>(
&self,
builder: TopicsBuilder<Uninit, E, B>
) -> <B as TopicsBuilderBackend<E>>::Outputwhere
E: Environment,
B: TopicsBuilderBackend<E>,

Guides event topic serialization using the given topics builder.