Module ink_env::call

Utilities to call or instantiate contracts on the chain.

Modules

state

Type states that tell what state of a instantiation argument has not yet been set properly for a valid construction.

utils

Utility types for the cross-contract calling API.

Structs

Call

The default call type for cross-contract calls. Performs a cross-contract call to callee with gas limit gas_limit, transferring transferred_value of currency.

CallBuilder

Builds up a cross contract call.

CallParams

The final parameters to the cross-contract call.

CreateBuilder

Builds up contract instantiations.

CreateParams

Builds up contract instantiations.

DelegateCall

The delegatecall call type. Performs a call with the given code hash.

ExecutionInput

The input data for a smart contract execution.

Selector

The function selector.

Traits

ConstructorReturnType

Represents any type that can be returned from an ink! constructor. The following contract implements the four different return type signatures implementing this trait:

FromAccountId

Contracts that can be constructed from an AccountId.

Functions

build_call

Returns a new CallBuilder to build up the parameters to a cross-contract call.

build_create

Returns a new CreateBuilder to build up the parameters to a cross-contract instantiation.