Module ink_env::call

Utilities to call or instantiate contracts on the chain.

Modules

state
Type states that tell what state of a instantiation argument has not yet been set properly for a valid construction.

utils
Utility types for the cross-contract calling API.

Structs

CallBuilder
Builds up a cross contract call.

CallParams
The final parameters to the cross-contract call.

CreateBuilder
Builds up contract instantiations.

CreateParams
Builds up contract instantiations.

ExecutionInput
The input data for a smart contract execution.

Selector
The function selector.

Traits

FromAccountId
Contracts that can be constructed from an AccountId.

Functions

build_call
Returns a new CallBuilder to build up the parameters to a cross-contract call.

build_create
Returns a new CreateBuilder to build up the parameters to a cross-contract instantiation.