Struct ink_env::call::CallBuilder

pub struct CallBuilder<E, CallType, Args, RetType>where
    E: Environment,{ /* private fields */ }

Builds up a cross contract call.

Implementations

impl<E, CallType, Args, RetType> CallBuilder<E, Unset<CallType>, Args, RetType>where E: Environment,

pub fn call_type<NewCallType>( self, call_type: NewCallType ) -> CallBuilder<E, Set<NewCallType>, Args, RetType>

The type of the call.

impl<E, CallType, Args, RetType> CallBuilder<E, CallType, Args, RetType>where E: Environment,

pub fn call_flags( self, call_flags: CallFlags ) -> CallBuilder<E, CallType, Args, RetType>

The flags used to change the behavior of the contract call.

impl<E, CallType, Args> CallBuilder<E, CallType, Args, Unset<ReturnType<()>>>where E: Environment,

pub fn returns<R>(self) -> CallBuilder<E, CallType, Args, Set<ReturnType<R>>>

Sets the type of the returned value upon the execution of the call.

Note

Either use .returns::<()> to signal that the call does not return a value or use .returns::<T> to signal that the call returns a value of type T.

impl<E, CallType, RetType> CallBuilder<E, CallType, Unset<ExecutionInput<EmptyArgumentList>>, RetType>where E: Environment,

pub fn exec_input<Args>( self, exec_input: ExecutionInput<Args> ) -> CallBuilder<E, CallType, Set<ExecutionInput<Args>>, RetType>

Sets the execution input to the given value.

impl<E, CallType, Args, RetType> CallBuilder<E, Unset<CallType>, Args, RetType>where E: Environment,

pub fn call( self, callee: E::AccountId ) -> CallBuilder<E, Set<Call<E>>, Args, RetType>

Prepares the CallBuilder for a cross-contract Call.

pub fn delegate( self, code_hash: E::Hash ) -> CallBuilder<E, Set<DelegateCall<E>>, Args, RetType>

Prepares the CallBuilder for a cross-contract DelegateCall.

impl<E, Args, RetType> CallBuilder<E, Set<Call<E>>, Args, RetType>where E: Environment,

pub fn gas_limit(self, gas_limit: Gas) -> Self

Sets the gas_limit for the current cross-contract call.

pub fn transferred_value(self, transferred_value: E::Balance) -> Self

Sets the transferred_value for the current cross-contract call.

impl<E, Args, RetType> CallBuilder<E, Set<DelegateCall<E>>, Args, RetType>where E: Environment,

pub fn code_hash(self, code_hash: E::Hash) -> Self

Sets the code_hash to perform a delegate call with.

impl<E, Args, RetType> CallBuilder<E, Set<Call<E>>, Set<ExecutionInput<Args>>, Set<ReturnType<RetType>>>where E: Environment,

pub fn params(self) -> CallParams<E, Call<E>, Args, RetType>

Finalizes the call builder to call a function.

impl<E, Args, RetType> CallBuilder<E, Set<DelegateCall<E>>, Set<ExecutionInput<Args>>, Set<ReturnType<RetType>>>where E: Environment,

pub fn params(self) -> CallParams<E, DelegateCall<E>, Args, RetType>

Finalizes the call builder to call a function.

impl<E, RetType> CallBuilder<E, Set<Call<E>>, Unset<ExecutionInput<EmptyArgumentList>>, Unset<RetType>>where E: Environment,

pub fn params(self) -> CallParams<E, Call<E>, EmptyArgumentList, ()>

Finalizes the call builder to call a function.

impl<E, RetType> CallBuilder<E, Set<DelegateCall<E>>, Unset<ExecutionInput<EmptyArgumentList>>, Unset<RetType>>where E: Environment,

pub fn params(self) -> CallParams<E, DelegateCall<E>, EmptyArgumentList, ()>

Finalizes the call builder to call a function.

impl<E> CallBuilder<E, Set<Call<E>>, Unset<ExecutionInput<EmptyArgumentList>>, Unset<ReturnType<()>>>where E: Environment,

pub fn invoke(self)

Invokes the cross-chain function call.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_invoke method instead.

pub fn try_invoke(self) -> Result<MessageResult<()>, Error>

Invokes the cross-chain function call.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E> CallBuilder<E, Set<DelegateCall<E>>, Unset<ExecutionInput<EmptyArgumentList>>, Unset<ReturnType<()>>>where E: Environment,

pub fn invoke(self)

Invokes the cross-chain function call using Delegate Call semantics.

Panics

This method panics if it encounters an ink::env::Error If you want to handle those use the try_invoke method instead.

pub fn try_invoke(self) -> Result<MessageResult<()>, Error>

Invokes the cross-chain function call using Delegate Call semantics.

Note

On failure this an ink::env::Error which can be handled by the caller.

impl<E, Args, R> CallBuilder<E, Set<Call<E>>, Set<ExecutionInput<Args>>, Set<ReturnType<R>>>where E: Environment, Args: Encode, R: Decode,

pub fn invoke(self) -> R

Invokes the cross-chain function call and returns the result.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_invoke method instead.

pub fn try_invoke(self) -> Result<MessageResult<R>, Error>

Invokes the cross-chain function call and returns the result.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E, Args, R> CallBuilder<E, Set<DelegateCall<E>>, Set<ExecutionInput<Args>>, Set<ReturnType<R>>>where E: Environment, Args: Encode, R: Decode,

pub fn invoke(self) -> R

Invokes the cross-chain function call using Delegate Call semantics and returns the result.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_invoke method instead.

pub fn try_invoke(self) -> Result<MessageResult<R>, Error>

Invokes the cross-chain function call using Delegate Call semantics and returns the result.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.