Struct ink_env::call::CallBuilder

pub struct CallBuilder<E, Callee, GasLimit, TransferredValue, Args, RetType> where
    E: Environment,  { /* private fields */ }

Builds up a cross contract call.

Implementations

impl<E, GasLimit, TransferredValue, Args, RetType> CallBuilder<E, Unset<E::AccountId>, GasLimit, TransferredValue, Args, RetType> where
E: Environment,

pub fn callee(
self,
callee: E::AccountId
) -> CallBuilder<E, Set<E::AccountId>, GasLimit, TransferredValue, Args, RetType>

Sets the called smart contract instance account ID to the given value.

impl<E, Callee, GasLimit, TransferredValue, Args, RetType> CallBuilder<E, Callee, GasLimit, TransferredValue, Args, RetType> where
E: Environment,

pub fn call_flags(
self,
call_flags: CallFlags
) -> CallBuilder<E, Callee, GasLimit, TransferredValue, Args, RetType>

The flags used to change the behavior of the contract call.

impl<E, Callee, TransferredValue, Args, RetType> CallBuilder<E, Callee, Unset<u64>, TransferredValue, Args, RetType> where
E: Environment,

pub fn gas_limit(
self,
gas_limit: u64
) -> CallBuilder<E, Callee, Set<u64>, TransferredValue, Args, RetType>

Sets the maximum allowed gas costs for the call.

impl<E, Callee, GasLimit, Args, RetType> CallBuilder<E, Callee, GasLimit, Unset<E::Balance>, Args, RetType> where
E: Environment,

pub fn transferred_value(
self,
transferred_value: E::Balance
) -> CallBuilder<E, Callee, GasLimit, Set<E::Balance>, Args, RetType>

Sets the value transferred upon the execution of the call.

impl<E, Callee, GasLimit, TransferredValue, Args> CallBuilder<E, Callee, GasLimit, TransferredValue, Args, Unset<ReturnType<()>>> where
E: Environment,

pub fn returns<R>(
self
) -> CallBuilder<E, Callee, GasLimit, TransferredValue, Args, Set<ReturnType<R>>>

Sets the type of the returned value upon the execution of the call.

Note

Either use .returns::<()> to signal that the call does not return a value or use .returns::<T> to signal that the call returns a value of type T.

impl<E, Callee, GasLimit, TransferredValue, RetType> CallBuilder<E, Callee, GasLimit, TransferredValue, Unset<ExecutionInput<EmptyArgumentList>>, RetType> where
E: Environment,

pub fn exec_input<Args>(
self,
exec_input: ExecutionInput<Args>
) -> CallBuilder<E, Callee, GasLimit, TransferredValue, Set<ExecutionInput<Args>>, RetType>

Sets the execution input to the given value.

impl<E, GasLimit, TransferredValue, Args, RetType> CallBuilder<E, Set<E::AccountId>, GasLimit, TransferredValue, Set<ExecutionInput<Args>>, Set<ReturnType<RetType>>> where
E: Environment,
GasLimit: Unwrap<Output = u64>,
TransferredValue: Unwrap<Output = E::Balance>,

pub fn params(self) -> CallParams<E, Args, RetType>

Finalizes the call builder to call a function.

impl<E, GasLimit, TransferredValue, RetType> CallBuilder<E, Set<E::AccountId>, GasLimit, TransferredValue, Unset<ExecutionInput<EmptyArgumentList>>, Unset<RetType>> where
E: Environment,
GasLimit: Unwrap<Output = u64>,
TransferredValue: Unwrap<Output = E::Balance>,

pub fn params(self) -> CallParams<E, EmptyArgumentList, ()>

Finalizes the call builder to call a function.

impl<E, GasLimit, TransferredValue> CallBuilder<E, Set<E::AccountId>, GasLimit, TransferredValue, Unset<ExecutionInput<EmptyArgumentList>>, Unset<ReturnType<()>>> where
E: Environment,
GasLimit: Unwrap<Output = u64>,
TransferredValue: Unwrap<Output = E::Balance>,

pub fn fire(self) -> Result<(), Error>

Invokes the cross-chain function call.

impl<E, GasLimit, TransferredValue, Args, R> CallBuilder<E, Set<E::AccountId>, GasLimit, TransferredValue, Set<ExecutionInput<Args>>, Set<ReturnType<R>>> where
E: Environment,
GasLimit: Unwrap<Output = u64>,
Args: Encode,
R: Decode,
TransferredValue: Unwrap<Output = E::Balance>,

pub fn fire(self) -> Result<R, Error>

Invokes the cross-chain function call and returns the result.