Struct ink_env::call::CallParams

pub struct CallParams<E, CallType, Args, R>
where
    E: Environment,
{ /* private fields */ }

The final parameters to the cross-contract call.

Implementations

impl<E, Args, R> CallParams<E, Call<E>, Args, R>
where E: Environment,

pub fn callee(&self) -> &E::AccountId

Returns the account ID of the called contract instance.

pub fn ref_time_limit(&self) -> u64

Returns the chosen ref time limit for the called contract execution.

pub fn proof_size_limit(&self) -> u64

Returns the chosen proof size limit for the called contract execution.

pub fn storage_deposit_limit(&self) -> Option<&E::Balance>

Returns the chosen storage deposit limit for the called contract execution.

pub fn transferred_value(&self) -> &E::Balance

Returns the transferred value for the called contract.

pub fn call_flags(&self) -> &CallFlags

Returns the call flags.

impl<E, Args, R> CallParams<E, Call<E>, Args, R>
where E: Environment, Args: Encode, R: Decode,

pub fn invoke(&self) -> R

Invokes the contract with the given built-up call parameters.

Returns the result of the contract execution.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_invoke method instead.

pub fn try_invoke(&self) -> Result<MessageResult<R>, Error>

Invokes the contract with the given built-up call parameters.

Returns the result of the contract execution.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E, Args, R> CallParams<E, CallV1<E>, Args, R>
where E: Environment,

pub fn callee(&self) -> &E::AccountId

Returns the account ID of the called contract instance.

pub fn gas_limit(&self) -> Gas

Returns the chosen gas limit for the called contract execution.

pub fn transferred_value(&self) -> &E::Balance

Returns the transferred value for the called contract.

pub fn call_flags(&self) -> &CallFlags

Returns the call flags.

impl<E, Args, R> CallParams<E, CallV1<E>, Args, R>
where E: Environment, Args: Encode, R: Decode,

pub fn invoke(&self) -> R

Invokes the contract with the given built-up call parameters.

Returns the result of the contract execution.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_invoke method instead.

pub fn try_invoke(&self) -> Result<MessageResult<R>, Error>

Invokes the contract with the given built-up call parameters.

Returns the result of the contract execution.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E, Args, R> CallParams<E, DelegateCall<E>, Args, R>
where E: Environment,

pub fn code_hash(&self) -> &E::Hash

Returns the code hash which we use to perform a delegate call.

pub fn call_flags(&self) -> &CallFlags

Returns the call flags.

impl<E, Args, R> CallParams<E, DelegateCall<E>, Args, R>
where E: Environment, Args: Encode, R: Decode,

pub fn invoke(&self) -> R

Invoke the contract using Delegate Call semantics with the given built-up call parameters.

Returns the result of the contract execution.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_invoke method instead.

pub fn try_invoke(&self) -> Result<MessageResult<R>, Error>

Invoke the contract using Delegate Call semantics with the given built-up call parameters.

Returns the result of the contract execution.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.

impl<E, CallType, Args, R> CallParams<E, CallType, Args, R>
where E: Environment,

pub fn exec_input(&self) -> &ExecutionInput<Args>

Returns the execution input.