ink_env::call

Struct CreateBuilder

pub struct CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>
where
    E: Environment,
{ /* private fields */ }

Builds up contract instantiations.

Implementations

impl<E, ContractRef, Limits, Args, RetType, Abi> CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>
where E: Environment,

pub fn code_hash( self, code_hash: H256, ) -> CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>

Sets the used code hash for the contract instantiation.

impl<E, ContractRef, Args, RetType, Abi> CreateBuilder<E, ContractRef, Set<LimitParamsV2>, Args, RetType, Abi>
where E: Environment,

pub fn ref_time_limit(self, ref_time_limit: u64) -> Self

Sets the ref_time_limit part of the weight limit for the contract instantiation.

pub fn proof_size_limit(self, proof_size_limit: u64) -> Self

Sets the proof_size_limit part of the weight limit for the contract instantiation.

pub fn storage_deposit_limit(self, storage_deposit_limit: U256) -> Self

Sets the storage_deposit_limit for the contract instantiation.

impl<E, ContractRef, Limits, Args, RetType, Abi> CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>
where E: Environment,

pub fn endowment( self, endowment: U256, ) -> CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, Limits, RetType, Abi> CreateBuilder<E, ContractRef, Limits, Unset<ExecutionInput<EmptyArgumentList<Abi>, Abi>>, RetType, Abi>
where E: Environment,

pub fn exec_input<Args>( self, exec_input: ExecutionInput<Args, Abi>, ) -> CreateBuilder<E, ContractRef, Limits, Set<ExecutionInput<Args, Abi>>, RetType, Abi>

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, Limits, Args, RetType, Abi> CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>
where E: Environment,

pub fn salt_bytes( self, salt: Option<[u8; 32]>, ) -> CreateBuilder<E, ContractRef, Limits, Args, RetType, Abi>

Sets the salt used for the execution of the call.

impl<E, ContractRef, Limits, Args, Abi> CreateBuilder<E, ContractRef, Limits, Args, Unset<ReturnType<()>>, Abi>
where E: Environment,

pub fn returns<R>( self, ) -> CreateBuilder<E, ContractRef, Limits, Args, Set<ReturnType<R>>, Abi>
where ContractRef: FromAddr, R: ConstructorReturnType<ContractRef, Abi>,

Sets the type of the returned value upon the execution of the constructor.

Note

Constructors are not able to return arbitrary values. Instead, a successful call to a constructor returns the address at which the contract was instantiated.

Therefore this must always be a reference (i.e. ContractRef) to the contract youโre trying to instantiate.

impl<E, ContractRef, Limits, Args, RetType, Abi> CreateBuilder<E, ContractRef, Set<Limits>, Set<ExecutionInput<Args, Abi>>, Set<ReturnType<RetType>>, Abi>
where E: Environment,

pub fn params(self) -> CreateParams<E, ContractRef, Limits, Args, RetType, Abi>

Finalizes the CreateBuilder, allowing it to instantiate a contract.

impl<E, ContractRef, Args, RetType, Abi> CreateBuilder<E, ContractRef, Set<LimitParamsV2>, Set<ExecutionInput<Args, Abi>>, Set<ReturnType<RetType>>, Abi>
where E: Environment, ContractRef: FromAddr + ContractReverseReference, <ContractRef as ContractReverseReference>::Type: ContractConstructorDecoder + ContractMessageDecoder, Args: EncodeArgsWith<Abi>, RetType: ConstructorReturnType<ContractRef, Abi>,

pub fn instantiate( self, ) -> <RetType as ConstructorReturnType<ContractRef, Abi>>::Output

todo check comment Instantiates the contract and returns its account ID back to the caller.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_instantiate method instead.

pub fn try_instantiate( self, ) -> Result<ConstructorResult<<RetType as ConstructorReturnType<ContractRef, Abi>>::Output>, Error>

todo check comment Instantiates the contract and returns its account ID back to the caller.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.