Struct ink_env::call::CreateBuilder

pub struct CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Args, Salt, RetType>where
    E: Environment,{ /* private fields */ }

Builds up contract instantiations.

Implementations

impl<E, ContractRef, GasLimit, Endowment, Args, Salt, RetType> CreateBuilder<E, ContractRef, Unset<E::Hash>, GasLimit, Endowment, Args, Salt, RetType>where
E: Environment,

pub fn code_hash(
self,
code_hash: E::Hash
) -> CreateBuilder<E, ContractRef, Set<E::Hash>, GasLimit, Endowment, Args, Salt, RetType>

Sets the used code hash for the contract instantiation.

impl<E, ContractRef, CodeHash, Endowment, Args, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, Unset<u64>, Endowment, Args, Salt, RetType>where
E: Environment,

pub fn gas_limit(
self,
gas_limit: u64
) -> CreateBuilder<E, ContractRef, CodeHash, Set<u64>, Endowment, Args, Salt, RetType>

Sets the maximum allowed gas costs for the contract instantiation.

impl<E, ContractRef, CodeHash, GasLimit, Args, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Unset<E::Balance>, Args, Salt, RetType>where
E: Environment,

pub fn endowment(
self,
endowment: E::Balance
) -> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Set<E::Balance>, Args, Salt, RetType>

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, CodeHash, GasLimit, Endowment, Salt, RetType> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Unset<ExecutionInput<EmptyArgumentList>>, Salt, RetType>where
E: Environment,

pub fn exec_input<Args>(
self,
exec_input: ExecutionInput<Args>
) -> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Set<ExecutionInput<Args>>, Salt, RetType>

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, CodeHash, GasLimit, Endowment, Args, RetType> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Args, Unset<Salt>, RetType>where
E: Environment,

pub fn salt_bytes<Salt>(
self,
salt: Salt
) -> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Args, Set<Salt>, RetType>where
Salt: AsRef<[u8]>,

Sets the value transferred upon the execution of the call.

impl<E, ContractRef, CodeHash, GasLimit, Endowment, Args, Salt> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Args, Salt, Unset<ReturnType<()>>>where
E: Environment,

pub fn returns<R>(
self
) -> CreateBuilder<E, ContractRef, CodeHash, GasLimit, Endowment, Args, Salt, Set<ReturnType<R>>>where
ContractRef: FromAccountId<E>,
R: ConstructorReturnType<ContractRef>,

Sets the type of the returned value upon the execution of the constructor.

Note

Constructors are not able to return arbitrary values. Instead a successful call to a constructor returns the address at which the contract was instantiated.

Therefore this must always be a reference (i.e ContractRef) to the contract you’re trying to instantiate.

impl<E, ContractRef, GasLimit, Args, Salt, RetType> CreateBuilder<E, ContractRef, Set<E::Hash>, GasLimit, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, Set<ReturnType<RetType>>>where
E: Environment,
GasLimit: Unwrap<Output = u64>,

pub fn params(self) -> CreateParams<E, ContractRef, Args, Salt, RetType>

Finalizes the create builder, allowing it to instantiate a contract.

impl<E, ContractRef, GasLimit, Args, Salt, RetType> CreateBuilder<E, ContractRef, Set<E::Hash>, GasLimit, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, Set<ReturnType<RetType>>>where
E: Environment,
ContractRef: FromAccountId<E>,
GasLimit: Unwrap<Output = u64>,
Args: Encode,
Salt: AsRef<[u8]>,
RetType: ConstructorReturnType<ContractRef>,

pub fn instantiate(
self
) -> <RetType as ConstructorReturnType<ContractRef>>::Output

Instantiates the contract and returns its account ID back to the caller.

Panics

This method panics if it encounters an ink::env::Error or an [ink::primitives::LangError][ink_primitives::LangError]. If you want to handle those use the try_instantiate method instead.

pub fn try_instantiate(
self
) -> Result<ConstructorResult<<RetType as ConstructorReturnType<ContractRef>>::Output>, Error>

Instantiates the contract and returns its account ID back to the caller.

Note

On failure this returns an outer ink::env::Error or inner [ink::primitives::LangError][ink_primitives::LangError], both of which can be handled by the caller.