Struct ink_env::call::CreateBuilder

pub struct CreateBuilder<E, CodeHash, GasLimit, Endowment, Args, Salt, R> where
    E: Environment,  { /* fields omitted */ }

Builds up contract instantiations.

Implementations

impl<E, GasLimit, Endowment, Args, Salt, R> CreateBuilder<E, Unset<E::Hash>, GasLimit, Endowment, Args, Salt, R> where
E: Environment,

pub fn code_hash(
self,
code_hash: E::Hash
) -> CreateBuilder<E, Set<E::Hash>, GasLimit, Endowment, Args, Salt, R>

Sets the used code hash for the contract instantiation.

impl<E, CodeHash, Endowment, Args, Salt, R> CreateBuilder<E, CodeHash, Unset<u64>, Endowment, Args, Salt, R> where
E: Environment,

pub fn gas_limit(
self,
gas_limit: u64
) -> CreateBuilder<E, CodeHash, Set<u64>, Endowment, Args, Salt, R>

Sets the maximum allowed gas costs for the contract instantiation.

impl<E, CodeHash, GasLimit, Args, Salt, R> CreateBuilder<E, CodeHash, GasLimit, Unset<E::Balance>, Args, Salt, R> where
E: Environment,

pub fn endowment(
self,
endowment: E::Balance
) -> CreateBuilder<E, CodeHash, GasLimit, Set<E::Balance>, Args, Salt, R>

Sets the value transferred upon the execution of the call.

impl<E, CodeHash, GasLimit, Endowment, Salt, R> CreateBuilder<E, CodeHash, GasLimit, Endowment, Unset<ExecutionInput<EmptyArgumentList>>, Salt, R> where
E: Environment,

pub fn exec_input<Args>(
self,
exec_input: ExecutionInput<Args>
) -> CreateBuilder<E, CodeHash, GasLimit, Endowment, Set<ExecutionInput<Args>>, Salt, R>

Sets the value transferred upon the execution of the call.

impl<E, CodeHash, GasLimit, Endowment, Args, R> CreateBuilder<E, CodeHash, GasLimit, Endowment, Args, Unset<Salt>, R> where
E: Environment,

pub fn salt_bytes<Salt>(
self,
salt: Salt
) -> CreateBuilder<E, CodeHash, GasLimit, Endowment, Args, Set<Salt>, R> where
Salt: AsRef<[u8]>,

Sets the value transferred upon the execution of the call.

impl<E, GasLimit, Args, Salt, R> CreateBuilder<E, Set<E::Hash>, GasLimit, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, R> where
E: Environment,
GasLimit: Unwrap<Output = u64>,

pub fn params(self) -> CreateParams<E, Args, Salt, R>

Sets the value transferred upon the execution of the call.

impl<E, GasLimit, Args, Salt, R> CreateBuilder<E, Set<E::Hash>, GasLimit, Set<E::Balance>, Set<ExecutionInput<Args>>, Set<Salt>, R> where
E: Environment,
GasLimit: Unwrap<Output = u64>,
Args: Encode,
Salt: AsRef<[u8]>,
R: FromAccountId<E>,

pub fn instantiate(self) -> Result<R, Error>

Instantiates the contract using the given instantiation parameters.