ink_env::call

Trait Executor

pub trait Executor<E: Environment> {
    type Error;

    // Required method
    fn exec<Args, Output, Abi>(
        &self,
        input: &ExecutionInput<Args, Abi>,
    ) -> Result<MessageResult<Output>, Self::Error>
       where Args: AbiEncodeWith<Abi>,
             Output: AbiDecodeWith<Abi> + DecodeMessageResult<Abi>;
}

Implemented in different environments to perform contract execution.

Required Associated Types

type Error

The type of the error that can be returned during execution.

Required Methods

fn exec<Args, Output, Abi>( &self, input: &ExecutionInput<Args, Abi>, ) -> Result<MessageResult<Output>, Self::Error>
where Args: AbiEncodeWith<Abi>, Output: AbiDecodeWith<Abi> + DecodeMessageResult<Abi>,

Perform the contract execution with the given input data, and return the result.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.