Module ink_env::call::utils

Utility types for the cross-contract calling API.

Structs

Argument
A single argument and its reference to a known value.

ArgumentList
An argument list.

ArgumentListEnd
The end of an argument list.

ReturnType
Represents a return type.

Set
A parameter that has been set to some value.

Unset
A parameter that has not been set, yet.

Traits

IndicateReturnType
Types that can be used in CallBuilder::returns to signal return type.

Unwrap
Implemented by Set and Unset in order to unwrap their value.

Type Definitions

ArgsList
Minor simplification of an argument list with a head and rest.

EmptyArgumentList
An empty argument list.