Module ink_env::chain_extension

Definitions and utilities for calling chain extension methods.

Users should not use these types and definitions directly but rather use the provided #[ink::chain_extension] procedural macro defined in the ink crate.

Modules

state

Type states of the chain extension method instance.

Structs

ChainExtensionMethod

A concrete instance of a chain extension method.

Traits

FromStatusCode

Implemented by error codes in order to construct them from status codes.

IsResultType

Extract Ok and Err variants from Result type.