Enum ink_env::Error

pub enum Error {
    Decode(Error),
    OffChain(OffChainError),
    CalleeTrapped,
    CalleeReverted,
    KeyNotFound,
    BelowSubsistenceThreshold,
    TransferFailed,
    NewContractNotFunded,
    CodeNotFound,
    NotCallable,
    UnknownError,
}

Errors that can be encountered upon environmental interaction.

Variants

Decode(Error)

Error upon decoding an encoded value.

OffChain(OffChainError)

An error that can only occur in the off-chain environment.

CalleeTrapped

The call to another contract has trapped.

CalleeReverted

The call to another contract has been reverted.

KeyNotFound

The queried contract storage entry is missing.

BelowSubsistenceThreshold

Transfer failed because it would have brought the sender’s total balance below the subsistence threshold.

TransferFailed

Transfer failed for other not further specified reason. Most probably reserved or locked balance of the sender that was preventing the transfer.

NewContractNotFunded

The newly created contract is below the subsistence threshold after executing its constructor so no usable contract instance will be created.

CodeNotFound

No code could be found at the supplied code hash.

NotCallable

The account that was called is either no contract (e.g. user account) or is a tombstone.

UnknownError

An unknown error has occurred.