Function ink_env::set_contract_storage

pub fn set_contract_storage<V>(key: &Key, value: &V) where
    V: Encode, 

Writes the value to the contract storage under the given key.

Panics

If the encode length of value exceeds the configured maximum value length of a storage entry.