Trait ink_env::AccountIdGuard

pub trait AccountIdGuard { }

A trait to enforce that a type should be an Environment::AccountId.

If you have an Environment which uses an Environment::AccountId type other than the ink! provided AccountId you will need to implement this trait for your Environment::AccountId concrete type.