Trait ink_env::Environment

pub trait Environment {
    type AccountId: 'static + Codec + Clone + PartialEq + Eq + Ord + AsRef<[u8]> + AsMut<[u8]>;
    type Balance: 'static + Codec + Copy + Clone + PartialEq + Eq + AtLeast32BitUnsigned + FromLittleEndian;
    type Hash: 'static + Codec + Copy + Clone + Clear + PartialEq + Eq + Ord + AsRef<[u8]> + AsMut<[u8]>;
    type Timestamp: 'static + Codec + Copy + Clone + PartialEq + Eq + AtLeast32BitUnsigned + FromLittleEndian;
    type BlockNumber: 'static + Codec + Copy + Clone + PartialEq + Eq + AtLeast32BitUnsigned + FromLittleEndian;
    type ChainExtension;

    const MAX_EVENT_TOPICS: usize;
}

The environmental types usable by contracts defined with ink!.

Required Associated Types

type AccountId: 'static + Codec + Clone + PartialEq + Eq + Ord + AsRef<[u8]> + AsMut<[u8]>

The account id type.

type Balance: 'static + Codec + Copy + Clone + PartialEq + Eq + AtLeast32BitUnsigned + FromLittleEndian

The type of balances.

type Hash: 'static + Codec + Copy + Clone + Clear + PartialEq + Eq + Ord + AsRef<[u8]> + AsMut<[u8]>

The type of hash.

type Timestamp: 'static + Codec + Copy + Clone + PartialEq + Eq + AtLeast32BitUnsigned + FromLittleEndian

The type of a timestamp.

type BlockNumber: 'static + Codec + Copy + Clone + PartialEq + Eq + AtLeast32BitUnsigned + FromLittleEndian

The type of block number.

type ChainExtension

The chain extension for the environment.

This is a type that is defined through the #[ink::chain_extension] procedural macro. For more information about usage and definition click this link.

Required Associated Constants

const MAX_EVENT_TOPICS: usize

The maximum number of supported event topics provided by the runtime.

The value must match the maximum number of supported event topics of the used runtime.