Module ink_env::types

Types for the default environment.

These are simple mirrored types from the default substrate FRAME configuration. Their interfaces and functionality might not be complete.

Users are required to provide their own type definitions and Environment implementations in order to write ink! contracts for other chain configurations.

Note

When authoring a contract, the concrete Environment are available via aliases generated by the lang macro. Therefore all functionality of the concrete types is accessible in the contract, not constrained by the required trait bounds.

Outside of the contract and its tests (e.g. in the offchain environment), where there is no knowledge of the concrete types, the functionality is restricted to the trait bounds on the Environment trait types.

Structs

AccountId
The default environment AccountId type.

Hash
The default environment Hash type.

Enums

DefaultEnvironment
The fundamental types of the default configuration.

Constants

_IMPL_TYPE_INFO_FOR_AccountId

_IMPL_TYPE_INFO_FOR_DefaultEnvironment

_IMPL_TYPE_INFO_FOR_Hash

Traits

Clear
The equivalent of Zero for hashes.

Environment
The environmental types usable by contracts defined with ink!.

Type Definitions

Balance
The default balance type.

BlockNumber
The default block number type.

Timestamp
The default timestamp type.